♻️ refacto routes
