+++ app/components/Navbar.tsx
@@ -18,7 +18,7 @@ export function Navbar() {
             </div>
             <ul className="flex flex-col gap-4 py-4 font-title desktop:flex-row desktop:items-center">
               <li>
-                <NavLink to="/qui-sommes-nous">
+                <NavLink to="/about">
                   <FiSmile />
                   Qui sommes-nous ?
                 </NavLink>
