+++ app/routes/legal.cgu.mdx
