+++ app/routes/media.1.mdx
@@ -81,11 +81,11 @@ Nous allons nous concentrer ici sur la première itération de cette discipline,
 jour : la partie **Revenue Operations, dédiée aux équipes Revenue (Sales, Marketing, Customer
 Success & Customer Care voire Finance).**
 
-## 💸 Le premier chapitre, Revenue Operations.
+## 💸 Le premier chapitre, Revenue Operations.
 
 ![Business ops revenue](/posts/1/2.png)
 
-### 🌱 De l’émergence initiale des Sales Ops…
+### 🌱 De l’émergence initiale des Sales Ops…
 
 On a vraiment vu émerger la notion d’opérations ces dernières années à travers le rôle de Sales Ops,
 l’équipe Sales étant la première à avoir bénéficié de ce type de profil dédié pour l’accompagner à
@@ -104,7 +104,7 @@ De fait, un Sales Ops est plus simple à défendre dans une organisation car son
 matérialiser par du CA supplémentaire. C’est pour cela qu’on l’a rapidement vu se déployer dans de
 nombreuses entreprises.
 
-### 🏡 … vers le besoin de plus d’ops
+### 🏡 … vers le besoin de plus d’ops
 
 Mais si les Sales Op sont essentiels, ils ne suffisent pas ! Si on optimise concrètement
 l’efficacité opérationnelle des équipes commerciales, on vend plus et mieux mais on ne répond à
@@ -139,7 +139,7 @@ adosser des Revenue Ops Manager, avec par exemple des Customer Success Ops Manag
 Mais pour rendre tout cela un peu moins théorique, nous allons essayer de résumer ici les missions
 des équipes Revenue Operations à nos yeux.
 
-### 🎫 La carte de visite d’une équipe Revenue Operations
+### 🎫 La carte de visite d’une équipe Revenue Operations
 
 ![Revenue operations position](/posts/1/4.png)
 
@@ -168,7 +168,7 @@ des équipes Revenue Operations à nos yeux.
 
 Cela se concrétise par **3 missions** claires :
 
-#### 🔎 Mission 1 : détecter et adresser les poches d’inefficacité des équipes
+#### 🔎 Mission 1 : détecter et adresser les poches d’inefficacité des équipes
 
 - Identifier les processus ou les outils qui manquent d’efficacité pour maximiser l’excellence
   opérationnelle des équipes Marketing, Sales, Service et Finance :
@@ -182,14 +182,14 @@ Cela se concrétise par **3 missions** claires :
   - Designer les nouveaux processus, les évolutions d’outils pour les supporter et définir le plan
     de déploiement.
 
-#### 🎯 Mission 2 : s’assurer de l’alignement de toutes les équipes sur les priorités
+#### 🎯 Mission 2 : s’assurer de l’alignement de toutes les équipes sur les priorités
 
 - Travailler avec les C-Level, les directeurs et la leadership team pour définir des OKRs
   trimestriels issues des projets identifiés en amont
 - Construire les plans de transformations avec des étapes réalistes et mesurables
 - Suivre la bonne mise en place de ce plan par les équipes
 
-#### **🚀** Mission 3 : soutenir la croissance de l’entreprise en déployant la vision opérationnelle
+#### **🚀** Mission 3 : soutenir la croissance de l’entreprise en déployant la vision opérationnelle
 
 - Construire et mettre à disposition de toutes les équipes Revenue les processus optimisés, les
   produits, les outils, les analyses, l’onboarding et le support nécessaire pour maximiser leur
