+++ app/routes/media.10.mdx
@@ -29,6 +29,7 @@ tags:
   - organisation
   - déploiement
 status: published
+slug: 10
 ---
 
 import { PostHeader } from '~/components/PostHeader';
