+++ app/routes/media.2.mdx
@@ -32,6 +32,7 @@ tags:
   - stratégie
   - process & outils
 status: published
+slug: 2
 ---
 
 import { PostHeader } from '~/components/PostHeader';
