+++ app/routes/media.3.mdx
@@ -32,6 +32,7 @@ tags:
   - organisation
   - l'ops
 status: published
+slug: 3
 ---
 
 import { PostHeader } from '~/components/PostHeader';
