+++ app/routes/media.4.mdx
@@ -32,6 +32,7 @@ tags:
   - stratégie
   - compensation
 status: published
+slug: 4
 ---
 
 import { PostHeader } from '~/components/PostHeader';
