+++ app/routes/media.5.mdx
@@ -28,6 +28,7 @@ tags:
   - déploiement
   - process & outils
 status: published
+slug: 5
 ---
 
 import { PostHeader } from '~/components/PostHeader';
