+++ app/routes/media.6.mdx
@@ -30,6 +30,7 @@ tags:
   - organisation
   - l'ops
 status: published
+slug: 6
 ---
 
 import { PostHeader } from '~/components/PostHeader';
