+++ app/routes/media.7.mdx
@@ -32,6 +32,7 @@ tags:
   - stratégie
   - compensation
 status: published
+slug: 7
 ---
 
 import { PostHeader } from '~/components/PostHeader';
