+++ app/routes/media.8.mdx
@@ -34,6 +34,7 @@ tags:
   - stratégie
   - compensation
 status: published
+slug: 8
 ---
 
 import { PostHeader } from '~/components/PostHeader';
