+++ app/routes/media.9.mdx
@@ -30,6 +30,7 @@ tags:
   - compensation
   - livre blanc
 status: published
+slug: 9
 ---
 
 import { PostHeader } from '~/components/PostHeader';
