--- app/routes/media/index.tsx
@@ -1,72 +1,0 @@
-import { json } from '@remix-run/node';
-import type { MetaFunction } from '@remix-run/react';
-import { NavLink, useLoaderData } from '@remix-run/react';
-
-import { PostCard } from '~/components/PostCard';
-import { postFromModule } from '~/utils/parsers';
-
-import * as post1 from './1.mdx';
-import * as post2 from './2.mdx';
-import * as post3 from './3.mdx';
-import * as post4 from './4.mdx';
-import * as post5 from './5.mdx';
-import * as post6 from './6.mdx';
-import * as post7 from './7.mdx';
-import * as post8 from './8.mdx';
-import * as post9 from './9.mdx';
-import * as post10 from './10.mdx';
-
-export async function loader() {
-  return json({
-    title: 'Ocobo • le Nouvel Ops',
-    description:
-      "Ocobo Le Nouvel Ops, c'est le média qui accompagne les Business Ops d'aujourd'hui et de demain. L’idée ? Vous éclairer sur diverses thématiques en vous partageant des articles sur les tendances du marché, les méthodologies, les bonnes pratiques & benchmarks qui gravitent autour de notre passionnant métier",
-    coverImage: 'https://ocobo.co/cover.png',
-    posts: [
-      postFromModule(post10),
-      postFromModule(post9),
-      postFromModule(post8),
-      postFromModule(post7),
-      postFromModule(post6),
-      postFromModule(post5),
-      postFromModule(post4),
-      postFromModule(post3),
-      postFromModule(post2),
-      postFromModule(post1),
-    ],
-  });
-}
-
-export const meta: MetaFunction<typeof loader> = ({ data, location }) => {
-  return [
-    { title: data?.title },
-    { name: 'description', content: data?.description },
-    { name: 'twitter:card', content: 'summary_large_image' },
-    { name: 'og:title', content: data?.title },
-    { name: 'og:type', content: 'article' },
-    { name: 'og:url', content: `https://ocobo.co${location.pathname}` },
-    { name: 'og:description', content: data?.description },
-    { name: 'og:image', content: data?.coverImage },
-  ];
-};
-
-export default function Index() {
-  const { posts } = useLoaderData<typeof loader>();
-
-  return (
-    <div className="not-prose min-h-screen">
-      <ul className="grid gap-8">
-        {posts.map((post, index) => (
-          <li
-            key={post.slug}
-            className="group overflow-hidden rounded-md bg-white shadow-lg dark:text-dark"
-          >
-            <NavLink to={post.slug} className="block">
-              <PostCard {...post.card} isHightlighted />
-            </NavLink>
-          </li>
-        ))}
-      </ul>
-    </div>
-  );
-}
