+++ app/routes/testimonial.leeway.mdx
@@ -1,0 +1,185 @@
+---
+meta:
+  - title: Leeway
+  - name: description
+    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
+  - name: og:title
+    content: Leeway
+  - name: og:type
+    content: article
+  - name: og:image
+    content: /testimonials/leeway/leeway-cover.jpg
+  - name: og:description
+    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
+  - name: og:url
+    content: https://ocobo.co/testimonial/leeway
+coverImage: /testimonials/leeway/leeway-cover.jpg
+date: 2023-08-14
+logo: /testimonials/leeway/leeway-logo.png
+tags:
+  - stratégie
+  - déploiement
+  - équipe & planification
+  - process & outils
+  - data & performance
+duration: 3 mois
+scope:
+  - Acquisition
+  - Demande
+  - Découverte
+guest: Thibaut Caoudal
+position: CRO
+tools:
+  - Hubspot
+  - Lemlist
+  - Gsheet
+title: Leeway
+slug: leeway
+quotes:
+  - Ocobo nous a aidé à prendre du recul sur notre organisation, nos process et nos outils pour
+    prendre les bonnes décisions business
+  - Ocobo a insufflé une vraie culture Ops transverse, pour aider Leeway à mieux suivre les
+    performances de ses équipes et optimiser ses résultats
+  - Ocobo a revu notre Ideal Customer Profile et en a construit un score de priorisation afin de
+    cibler les entreprises les plus pertinentes à assigner aux commerciaux avant vente
+---
+
+import { TestimonialHeader } from '~/components/TestimonialHeader';
+import { Callout } from '~/components/Callout';
+
+export default function Layout({ children }) {
+  return <div className="grid-cols-12 gap-4 desktop:grid">{children}</div>;
+}
+
+<div className="col-span-12">
+  <TestimonialHeader
+    title={attributes.meta.find((m) => m.title).title}
+    guest={attributes.guest}
+    position={attributes.position}
+    duration={attributes.duration}
+    coverImage={attributes.coverImage}
+    logo={attributes.logo}
+    tags={attributes.tags}
+    scope={attributes.scope}
+    tools={attributes.tools}
+    description={attributes.meta.find((m) => m.name === 'description').content}
+  />
+</div>
+
+<div className="prose col-span-6">
+
+## ✉️ La mission
+
+L’objectif de la mission était d’accompagner Leeway dans l’**industrialisation du canal
+d’acquisition outbound** afin de les aider à mieux cibler les prospects et ainsi créer des
+fondations solides pour la phase de croissance qui s’annonce.
+
+La mission s’est déroulée en 3 phases :
+
+**Phase 1 - Optimiser les process Growth Ops** pour gagner en temps et en efficacité dans la
+génération de comptes à prospecter
+
+<div className="relief">
+
+🎯 Objectifs :
+
+1. Réaliser un audit avec recommandation des process growth en place ;
+2. Communiquer sur les bonnes pratiques du marché ;
+3. Coacher la ressource Growth Ops interne.
+
+</div>
+
+**Phase 2 - Restructurer la stratégie outbound** de la phase d’allocation des comptes à la
+génération du bon nombre d’opportunités pour chaque SDR et AE (redéfinir l’audience, mieux la
+prioriser, standardiser le cycle de vente et les process).
+
+<div className="relief">
+
+🎯 Objectifs :
+
+1. Réaliser un audit avec recommandation des processus commerciaux outbound ;
+2. Identifier les principaux sujets d’optimisation à lancer et prioriser ;
+3. Définir de l’organisation cible, des nouveaux processus, de l’architecture des outils, la
+   nouvelle segmentation, les automatisations et les KPIs à suivre.
+
+</div>
+
+**Phase 3 - Déployer dans les outils (Hubspot) et former les équipes**
+
+<div className="relief">
+
+🎯 Objectifs :
+
+1. Migration et construction des workflows dans Hubspot incluant une phase de test ;
+2. Reprise de la donnée historique ;
+3. Documentation et animation de la formation des équipes.
+
+</div>
+
+<Callout>
+
+### 📚 Livrables de la mission
+
+- **Stratégie Growth** avec la clarification de l’ICP (Ideal Customer Profile - les meilleurs
+  prospects à contacter), définition et mise en place d’un score de priorisation et d’intentions
+  afin de sélectionner les meilleurs comptes à assigner aux SDRs;
+- **Définition** et mise en place dans Hubspot du **nouveau cycle de vente** avec les étapes
+  principales (entreprise, contact, transaction), leurs taux de conversion et les principaux KPIs
+  associés;
+- **Automatisation** des étapes du cycle de vente dans Hubspot entre les objets;
+- **Mise en place d’outils** pour faire gagner en **productivité** les équipes commerciales : vues
+  filtrées, séquences de contact, KPIs, dashboard de suivi, etc.
+- **Re-définition** et implémentation d’une nouvelle **méthode de qualification** grâce au MEDDIC;
+- **Documentation** et **formation** des équipes on-site;
+
+</Callout>
+
+</div>
+
+<div className="col-span-1" />
+
+<div className="prose col-span-5">
+
+## 🎙️ Le retour d’expérience
+
+### Que fait Leeway ?
+
+> Leeway est une solution SaaS de contract management permettant aux entreprises de gérer l’ensemble
+> de leur cycle de vie contractuel de manière collaborative et centralisée.
+
+### À quel moment et pourquoi avoir fait appel à Ocobo ?
+
+> Nous avons fait appel à Ocobo pour préparer la phase de “scale” et faire grandir l’équipe
+> commerciale. Nous avions besoin de process et données fiables pour attaquer la prochaine étape de
+> la vie de Leeway et accélérer notre croissance.
+
+### Quelle problématique Ocobo est venue adresser et comment l’ont ils approché ?
+
+> À l’origine, nous avons sollicité Ocobo sur des **problématiques Outbound.** Nous avions besoin de
+> gagner en productivité sur l’ensemble du process commercial pour améliorer les résultats de
+> l’équipe. D’une part, la création des listes de comptes à allouer aux commerciaux était très
+> fastidieuse pour l’équipe Growth, et d’autre part, les commerciaux passaient trop de temps sur des
+> tâches à faible valeur ajoutée. **Ocobo nous a aidé à prendre du recul sur notre organisation, nos
+> process et nos outils pour prendre les bonnes décisions business.**
+>
+> Concrètement, **Ocobo a revu notre ICP** et en a **construit un account score** afin de cibler les
+> entreprises les plus pertinentes à assigner aux SDR. Puis dans une optique de construire les
+> fondations solides pour préparer la forte croissance de nos équipes de vente (SDR et Account
+> Executive), **Ocobo a redéfini notre cycle de vente de la génération des leads jusqu’à la
+> signature des opportunités**, les grandes étapes et a mis en place les bons outils afin de suivre
+> et mieux monitorer la performance des équipes (mise en place de nouveaux KPIs, dashboard, etc..).
+
+### Qu’est ce que Ocobo a apporté à Leeway ?
+
+> **Ocobo a insufflé une vraie culture Ops transverse,** pour aider Leeway à adresser les problèmes
+> liés au Revenue dans leur ensemble : analyse de données sur l’existant pour prendre les bonnes
+> décisions, approche ROIste sur les résultats, solution déployée impactant les process et les
+> outils, mais aussi optimisant l’organisation et améliorant la fiabilité et l’accès aux KPIs
+> importants.
+
+### Quelles sont les prochaines étapes pour Leeway ?
+
+> Aujourd’hui, grâce à Ocobo nous sommes prêts à faire grandir les équipes commerciales de manière
+> sereine avec les bons outils pour piloter la performance commerciale de l’entreprise.
+
+</div>
