+++ app/routes/testimonial.qare.mdx
@@ -1,0 +1,147 @@
+---
+meta:
+  - title: Qare
+  - name: description
+    content: Refonte du parcours de contact et de l’organisation du Service Client
+  - name: og:title
+    content: Qare
+  - name: og:type
+    content: article
+  - name: og:image
+    content: /testimonials/qare/qare-cover.jpg
+  - name: og:description
+    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
+  - name: og:url
+    content: https://ocobo.co/testimonial/qare
+coverImage: /testimonials/qare/qare-cover.jpeg
+date: 2023-20-09
+logo: /testimonials/qare/qare-logo.png
+tags:
+  - déploiement
+  - process & outils
+duration: 1 mois
+scope:
+  - Customer Service
+  - Expansion
+guest: Sébastien Panchaud
+position: COO
+tools:
+  - Aircall
+  - Salesforce
+  - Zendesk
+title: Qare
+slug: qare
+quotes:
+  - Nous avons fait appel à Ocobo pour nous permettre de reparamétrer le nouveau parcours afin de
+    pouvoir gérer les flux sans baisser la qualité de notre service
+  - Ocobo nous a apporté une nouvelle façon d’aborder la solution en prenant en compte le besoin,
+    l’organisation et l’expérience client dans son ensemble
+  - Ocobo nous a permis de mieux nous servir des capacités de Zendesk, et nous ont suggéré une
+    nouvelle organisation pour itérer et mieux gérer le partage d’information avec les équipes
+    Produit
+---
+
+import { TestimonialHeader } from '~/components/TestimonialHeader';
+import { Callout } from '~/components/Callout';
+
+export default function Layout({ children }) {
+  return <div className="grid-cols-12 gap-4 desktop:grid">{children}</div>;
+}
+
+<div className="col-span-12">
+  <TestimonialHeader
+    title={attributes.meta.find((m) => m.title).title}
+    guest={attributes.guest}
+    position={attributes.position}
+    duration={attributes.duration}
+    coverImage={attributes.coverImage}
+    logo={attributes.logo}
+    tags={attributes.tags}
+    scope={attributes.scope}
+    tools={attributes.tools}
+    description={attributes.meta.find((m) => m.name === 'description').content}
+  />
+</div>
+
+<div className="prose col-span-6">
+
+## ✉️ La mission
+
+**L’objectif de cette mission était** d’accompagner Qare dans l’optimisation de la gestion de ses
+demandes clients afin de diminuer le volume de tickets tout en proposant la meilleure expérience
+utilisateur possible.
+
+Pour cela, nous avons travaillé sur 4 axes différents :
+
+1. **Canal Chat prioritaire** : analyser, prioriser et optimiser l’arbre de décision du **chatbot
+   Zendesk** de gestion des demandes
+2. **Gestion des tickets** : implémenter dans Zendesk des processus clairs et automatisés afin
+   d’optimiser le travail des agents
+3. **FAQ** : optimiser le helpcenter afin de réduire le volume de demandes à faible valeur ajoutée
+   donc optimiser les coûts
+4. **Organisation** : analyser l’organisation de l’équipe sur l’ensemble des canaux pour aider à
+   l’efficacité opérationnelle
+
+ <Callout>
+
+### 📚 Livrables de la mission
+
+- **Chatbot :** Revue du bot de Zendesk dans l’objectif d’exploiter ses capacités. Création de
+  nouveaux arbres de décisions optimisés afin de permettre au bot de comprendre des “phrases”
+  fournies au préalable, et de déclencher une réponse structurée en fonction de la demande
+  utilisateur (un article de la FAQ ou autre) plutôt que de contacter un agent directement.
+- **Gestion des tickets :** revue de la structuration de la catégorisation des tickets pour la
+  rendre plus en adéquation avec la nouvelle arborescence du bot, de la FAQ et plus digeste pour les
+  équipes produit.
+- **FAQ** : Revue du design du helpcenter et de l’arborescence des articles pour les rendre plus
+  digestes et actionnables.
+- **Organisation** : proposition d’organisation optimisée pour une meilleure gestion des ressources,
+  un lien plus fort avec les équipes Produit et plus d’autonomie et de flexibilité dans
+  l’administration de Zendesk.
+
+</Callout>
+
+</div>
+
+<div className="col-span-1" />
+<div className="prose col-span-5">
+
+## 🎙️ Le retour d’expérience
+
+### Que fait Qare ?
+
+> Qare est une plateforme de téléconsultation permettant aux patients d’avoir rapidement accès à des
+> soins de qualité.
+
+### À quel moment et pourquoi avoir fait appel à Ocobo ?
+
+> Nous avons changé d’outil de traitement des chats (Intercom > Zendesk) au sein de notre service
+> client et la migration a généré une forte augmentation du nombre de tickets que nous avions de la
+> peine à gérer. Nous avons fait appel à Ocobo pour nous permettre de reparamétrer le nouveau
+> parcours afin de pouvoir gérer les flux sans baisser la qualité de notre service.
+
+### Quelle problématique Ocobo est venu adresser et comment l’ont ils approché ?
+
+> Ocobo a principalement revu et repensé le parcours utilisateur qui permet d’avoir accès à un agent
+> du service client. Ils se sont basés sur leur expérience dans le domaine et leur connaissance de
+> l’outil Zendesk afin de proposer une nouvelle approche en: 1/ Analysant la typologie de demandes
+> entrantes 2/ Repensant les catégories et le contenu du helpcenter en fonction 3/ Déployant des
+> propositions d’articles et quelques réponses automatiquement dans le chat afin de réduire les
+> demandes
+>
+> Ils ont également réalisé un audit de notre organisation.
+
+### Qu’est ce que Ocobo a apporté à Qare ?
+
+> Une nouvelle façon d’aborder la solution en prenant en compte le besoin, l’organisation et
+> l’expérience client dans son ensemble. Ils nous ont permis de mieux nous servir des capacités de
+> l’outil, et nous ont suggéré une nouvelle organisation pour itérer et mieux gérer le partage
+> d’information avec les équipes Produit.
+
+### Quelles sont les prochaines étapes pour Qare concernant son Service Client ?
+
+> Nous continuons d’itérer le parcours mis en place et d’adapter la catégorisation des tickets et de
+> notre helpcenter en fonction des évolutions. Nous mettons aussi actuellement la nouvelle
+> organisation proposée par Ocobo en place. Merci à eux !
+
+</div>
