--- app/routes/testimonial/leeway/deliverables.mdx
@@ -1,13 +1,0 @@
-### 📚 Livrables de la mission
-
-- **Stratégie Growth** avec la clarification de l’ICP (Ideal Customer Profile - les meilleurs
-  prospects à contacter), définition et mise en place d’un score de priorisation et d’intentions
-  afin de sélectionner les meilleurs comptes à assigner aux SDRs;
-- **Définition** et mise en place dans Hubspot du **nouveau cycle de vente** avec les étapes
-  principales (entreprise, contact, transaction), leurs taux de conversion et les principaux KPIs
-  associés;
-- **Automatisation** des étapes du cycle de vente dans Hubspot entre les objets;
-- **Mise en place d’outils** pour faire gagner en **productivité** les équipes commerciales : vues
-  filtrées, séquences de contact, KPIs, dashboard de suivi, etc.
-- **Re-définition** et implémentation d’une nouvelle **méthode de qualification** grâce au MEDDIC;
-- **Documentation** et **formation** des équipes on-site;
