--- app/routes/testimonial/leeway/index.mdx
@@ -1,81 +1,0 @@
----
-meta:
-  - title: Leeway
-  - name: description
-    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
-  - name: og:title
-    content: Leeway
-  - name: og:type
-    content: article
-  - name: og:image
-    content: /testimonials/leeway/leeway-cover.jpg
-  - name: og:description
-    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
-  - name: og:url
-    content: https://ocobo.co/testimonial/leeway
-coverImage: /testimonials/leeway/leeway-cover.jpg
-date: 2023-08-14
-logo: /testimonials/leeway/leeway-logo.png
-tags:
-  - stratégie
-  - déploiement
-  - équipe & planification
-  - process & outils
-  - data & performance
-duration: 3 mois
-scope:
-  - Acquisition
-  - Demande
-  - Découverte
-guest: Thibaut Caoudal
-position: CRO
-tools:
-  - Hubspot
-  - Lemlist
-  - Gsheet
-title: Leeway
-slug: leeway
-quotes:
-  - Ocobo nous a aidé à prendre du recul sur notre organisation, nos process et nos outils pour
-    prendre les bonnes décisions business
-  - Ocobo a insufflé une vraie culture Ops transverse, pour aider Leeway à mieux suivre les
-    performances de ses équipes et optimiser ses résultats
-  - Ocobo a revu notre Ideal Customer Profile et en a construit un score de priorisation afin de
-    cibler les entreprises les plus pertinentes à assigner aux commerciaux avant vente
----
-
-import { TestimonialHeader } from '~/components/TestimonialHeader';
-import { Callout } from '~/components/Callout';
-import Missions from './mission.mdx';
-import Interview from './interview.mdx';
-import Deliverables from './deliverables.mdx';
-
-export default function Layout({ children }) {
-  return <div className="grid-cols-12 gap-4 desktop:grid">{children}</div>;
-}
-
-<div className="col-span-12">
-  <TestimonialHeader
-    title={attributes.meta.find((m) => m.title).title}
-    guest={attributes.guest}
-    position={attributes.position}
-    duration={attributes.duration}
-    coverImage={attributes.coverImage}
-    logo={attributes.logo}
-    tags={attributes.tags}
-    scope={attributes.scope}
-    tools={attributes.tools}
-    description={attributes.meta.find((m) => m.name === 'description').content}
-  />
-</div>
-
-<div className="prose col-span-6">
-  <Missions />
-  <Callout>
-    <Deliverables />
-  </Callout>
-</div>
-<div className="col-span-1" />
-<div className="prose col-span-5">
-  <Interview cover={attributes.coverImage} />
-</div>
