--- app/routes/testimonial/leeway/interview.mdx
@@ -1,41 +1,0 @@
-## 🎙️ Le retour d’expérience
-
-### Que fait Leeway ?
-
-> Leeway est une solution SaaS de contract management permettant aux entreprises de gérer l’ensemble
-> de leur cycle de vie contractuel de manière collaborative et centralisée.
-
-### À quel moment et pourquoi avoir fait appel à Ocobo ?
-
-> Nous avons fait appel à Ocobo pour préparer la phase de “scale” et faire grandir l’équipe
-> commerciale. Nous avions besoin de process et données fiables pour attaquer la prochaine étape de
-> la vie de Leeway et accélérer notre croissance.
-
-### Quelle problématique Ocobo est venue adresser et comment l’ont ils approché ?
-
-> À l’origine, nous avons sollicité Ocobo sur des **problématiques Outbound.** Nous avions besoin de
-> gagner en productivité sur l’ensemble du process commercial pour améliorer les résultats de
-> l’équipe. D’une part, la création des listes de comptes à allouer aux commerciaux était très
-> fastidieuse pour l’équipe Growth, et d’autre part, les commerciaux passaient trop de temps sur des
-> tâches à faible valeur ajoutée. **Ocobo nous a aidé à prendre du recul sur notre organisation, nos
-> process et nos outils pour prendre les bonnes décisions business.**
->
-> Concrètement, **Ocobo a revu notre ICP** et en a **construit un account score** afin de cibler les
-> entreprises les plus pertinentes à assigner aux SDR. Puis dans une optique de construire les
-> fondations solides pour préparer la forte croissance de nos équipes de vente (SDR et Account
-> Executive), **Ocobo a redéfini notre cycle de vente de la génération des leads jusqu’à la
-> signature des opportunités**, les grandes étapes et a mis en place les bons outils afin de suivre
-> et mieux monitorer la performance des équipes (mise en place de nouveaux KPIs, dashboard, etc..).
-
-### Qu’est ce que Ocobo a apporté à Leeway ?
-
-> **Ocobo a insufflé une vraie culture Ops transverse,** pour aider Leeway à adresser les problèmes
-> liés au Revenue dans leur ensemble : analyse de données sur l’existant pour prendre les bonnes
-> décisions, approche ROIste sur les résultats, solution déployée impactant les process et les
-> outils, mais aussi optimisant l’organisation et améliorant la fiabilité et l’accès aux KPIs
-> importants.
-
-### Quelles sont les prochaines étapes pour Leeway ?
-
-> Aujourd’hui, grâce à Ocobo nous sommes prêts à faire grandir les équipes commerciales de manière
-> sereine avec les bons outils pour piloter la performance commerciale de l’entreprise.
