--- app/routes/testimonial/leeway/mission.mdx
@@ -1,47 +1,0 @@
-## ✉️ La mission
-
-L’objectif de la mission était d’accompagner Leeway dans l’**industrialisation du canal
-d’acquisition outbound** afin de les aider à mieux cibler les prospects et ainsi créer des
-fondations solides pour la phase de croissance qui s’annonce.
-
-La mission s’est déroulée en 3 phases :
-
-**Phase 1 - Optimiser les process Growth Ops** pour gagner en temps et en efficacité dans la
-génération de comptes à prospecter
-
-<div className="relief">
-
-🎯 Objectifs :
-
-1. Réaliser un audit avec recommandation des process growth en place ;
-2. Communiquer sur les bonnes pratiques du marché ;
-3. Coacher la ressource Growth Ops interne.
-
-</div>
-
-**Phase 2 - Restructurer la stratégie outbound** de la phase d’allocation des comptes à la
-génération du bon nombre d’opportunités pour chaque SDR et AE (redéfinir l’audience, mieux la
-prioriser, standardiser le cycle de vente et les process).
-
-<div className="relief">
-
-🎯 Objectifs :
-
-1. Réaliser un audit avec recommandation des processus commerciaux outbound ;
-2. Identifier les principaux sujets d’optimisation à lancer et prioriser ;
-3. Définir de l’organisation cible, des nouveaux processus, de l’architecture des outils, la
-   nouvelle segmentation, les automatisations et les KPIs à suivre.
-
-</div>
-
-**Phase 3 - Déployer dans les outils (Hubspot) et former les équipes**
-
-<div className="relief">
-
-🎯 Objectifs :
-
-1. Migration et construction des workflows dans Hubspot incluant une phase de test ;
-2. Reprise de la donnée historique ;
-3. Documentation et animation de la formation des équipes.
-
-</div>
