--- app/routes/testimonial/qare/deliverables.mdx
@@ -1,14 +1,0 @@
-### 📚 Livrables de la mission
-
-- **Chatbot :** Revue du bot de Zendesk dans l’objectif d’exploiter ses capacités. Création de
-  nouveaux arbres de décisions optimisés afin de permettre au bot de comprendre des “phrases”
-  fournies au préalable, et de déclencher une réponse structurée en fonction de la demande
-  utilisateur (un article de la FAQ ou autre) plutôt que de contacter un agent directement.
-- **Gestion des tickets :** revue de la structuration de la catégorisation des tickets pour la
-  rendre plus en adéquation avec la nouvelle arborescence du bot, de la FAQ et plus digeste pour les
-  équipes produit.
-- **FAQ** : Revue du design du helpcenter et de l’arborescence des articles pour les rendre plus
-  digestes et actionnables.
-- **Organisation** : proposition d’organisation optimisée pour une meilleure gestion des ressources,
-  un lien plus fort avec les équipes Produit et plus d’autonomie et de flexibilité dans
-  l’administration de Zendesk.
