--- app/routes/testimonial/qare/index.mdx
@@ -1,78 +1,0 @@
----
-meta:
-  - title: Qare
-  - name: description
-    content: Refonte du parcours de contact et de l’organisation du Service Client
-  - name: og:title
-    content: Qare
-  - name: og:type
-    content: article
-  - name: og:image
-    content: /testimonials/qare/qare-cover.jpg
-  - name: og:description
-    content: Optimisation des process d’acquisition outbound (Growth, Demande et Découverte)
-  - name: og:url
-    content: https://ocobo.co/testimonial/qare
-coverImage: /testimonials/qare/qare-cover.jpeg
-date: 2023-20-09
-logo: /testimonials/qare/qare-logo.png
-tags:
-  - déploiement
-  - process & outils
-duration: 1 mois
-scope:
-  - Customer Service
-  - Expansion
-guest: Sébastien Panchaud
-position: COO
-tools:
-  - Aircall
-  - Salesforce
-  - Zendesk
-title: Qare
-slug: qare
-quotes:
-  - Nous avons fait appel à Ocobo pour nous permettre de reparamétrer le nouveau parcours afin de
-    pouvoir gérer les flux sans baisser la qualité de notre service
-  - Ocobo nous a apporté une nouvelle façon d’aborder la solution en prenant en compte le besoin,
-    l’organisation et l’expérience client dans son ensemble
-  - Ocobo nous a permis de mieux nous servir des capacités de Zendesk, et nous ont suggéré une
-    nouvelle organisation pour itérer et mieux gérer le partage d’information avec les équipes
-    Produit
----
-
-import { TestimonialHeader } from '~/components/TestimonialHeader';
-import { Callout } from '~/components/Callout';
-import Missions from './mission.mdx';
-import Interview from './interview.mdx';
-import Deliverables from './deliverables.mdx';
-
-export default function Layout({ children }) {
-  return <div className="grid-cols-12 gap-4 desktop:grid">{children}</div>;
-}
-
-<div className="col-span-12">
-  <TestimonialHeader
-    title={attributes.meta.find((m) => m.title).title}
-    guest={attributes.guest}
-    position={attributes.position}
-    duration={attributes.duration}
-    coverImage={attributes.coverImage}
-    logo={attributes.logo}
-    tags={attributes.tags}
-    scope={attributes.scope}
-    tools={attributes.tools}
-    description={attributes.meta.find((m) => m.name === 'description').content}
-  />
-</div>
-
-<div className="prose col-span-6">
-  <Missions />
-  <Callout>
-    <Deliverables />
-  </Callout>
-</div>
-<div className="col-span-1" />
-<div className="prose col-span-5">
-  <Interview cover={attributes.coverImage} />
-</div>
