--- app/routes/testimonial/qare/interview.mdx
@@ -1,37 +1,0 @@
-## 🎙️ Le retour d’expérience
-
-### Que fait Qare ?
-
-> Qare est une plateforme de téléconsultation permettant aux patients d’avoir rapidement accès à des
-> soins de qualité.
-
-### À quel moment et pourquoi avoir fait appel à Ocobo ?
-
-> Nous avons changé d’outil de traitement des chats (Intercom > Zendesk) au sein de notre service
-> client et la migration a généré une forte augmentation du nombre de tickets que nous avions de la
-> peine à gérer. Nous avons fait appel à Ocobo pour nous permettre de reparamétrer le nouveau
-> parcours afin de pouvoir gérer les flux sans baisser la qualité de notre service.
-
-### Quelle problématique Ocobo est venu adresser et comment l’ont ils approché ?
-
-> Ocobo a principalement revu et repensé le parcours utilisateur qui permet d’avoir accès à un agent
-> du service client. Ils se sont basés sur leur expérience dans le domaine et leur connaissance de
-> l’outil Zendesk afin de proposer une nouvelle approche en: 1/ Analysant la typologie de demandes
-> entrantes 2/ Repensant les catégories et le contenu du helpcenter en fonction 3/ Déployant des
-> propositions d’articles et quelques réponses automatiquement dans le chat afin de réduire les
-> demandes
->
-> Ils ont également réalisé un audit de notre organisation.
-
-### Qu’est ce que Ocobo a apporté à Qare ?
-
-> Une nouvelle façon d’aborder la solution en prenant en compte le besoin, l’organisation et
-> l’expérience client dans son ensemble. Ils nous ont permis de mieux nous servir des capacités de
-> l’outil, et nous ont suggéré une nouvelle organisation pour itérer et mieux gérer le partage
-> d’information avec les équipes Produit.
-
-### Quelles sont les prochaines étapes pour Qare concernant son Service Client ?
-
-> Nous continuons d’itérer le parcours mis en place et d’adapter la catégorisation des tickets et de
-> notre helpcenter en fonction des évolutions. Nous mettons aussi actuellement la nouvelle
-> organisation proposée par Ocobo en place. Merci à eux !
