--- app/routes/testimonial/qare/mission.mdx
@@ -1,16 +1,0 @@
-## ✉️ La mission
-
-**L’objectif de cette mission était** d’accompagner Qare dans l’optimisation de la gestion de ses
-demandes clients afin de diminuer le volume de tickets tout en proposant la meilleure expérience
-utilisateur possible.
-
-Pour cela, nous avons travaillé sur 4 axes différents :
-
-1. **Canal Chat prioritaire** : analyser, prioriser et optimiser l’arbre de décision du **chatbot
-   Zendesk** de gestion des demandes
-2. **Gestion des tickets** : implémenter dans Zendesk des processus clairs et automatisés afin
-   d’optimiser le travail des agents
-3. **FAQ** : optimiser le helpcenter afin de réduire le volume de demandes à faible valeur ajoutée
-   donc optimiser les coûts
-4. **Organisation** : analyser l’organisation de l’équipe sur l’ensemble des canaux pour aider à
-   l’efficacité opérationnelle
