+++ remix.config.js
@@ -1,10 +1,4 @@
-const { createRoutesFromFolders } = require('@remix-run/v1-route-convention');
 module.exports = {
   ignoredRouteFiles: ['**/.*'],
   serverModuleFormat: 'cjs',
-
-  routes(defineRoutes) {
-    // uses the v1 convention, works in v1.15+ and v2
-    return createRoutesFromFolders(defineRoutes);
-  },
 };
